requirements.txt
scikit-learn==1.6.1
streamlit==1.41.1
matplotlib==3.10.0
seaborn==0.13.2
python-dotenv==0.21.0
surprise-scikit==0.1
wordcloud==1.9.4
plotly==5.24.1
networkx==3.4.2